+++ backend/auth_service/internal/pkg/authjwt/signature/signature_PS256.go
@@ -1,4 +1,4 @@
-// Package vaulttoken реализует алгоритмы подписи и верификации токенов JWT с использованием Hashicorp Vault.
+// Package signature реализует алгоритмы подписи и верификации токенов JWT с использованием Hashicorp Vault.
 // Реализована подпись с использованием алгоритма RSA-PSS (PS256), где приватный ключ хранится в Vault,
 // а публичный ключ извлекается посредством интерфейса.
 package signature
